% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran-ls.R
\name{cran_db_refresh}
\alias{cran_db_refresh}
\title{Refresh a cached CRAN packages DB.}
\usage{
cran_db_refresh()
}
\value{
logical TRUE, or FALSE if no such cache existed
}
\description{
The vsils function \code{cran_package_db()} is \emph{memoized}, meaning that its output
is cached by the memoise package, with a timeout every hour. Manually bust this
cache by running \code{cran_db_refresh()}.
}
\examples{
\dontrun{
cran_db_refresh()
}
}
