% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{cran_package_db}
\alias{cran_package_db}
\title{Get CRAN package database.}
\usage{
cran_package_db()
}
\description{
This is a memoized (cached) copy of \code{tools::CRAN_package_db()}, it refreshes once an
hour.
}
\details{
Force refresh with \code{cran_db_refresh()}.
}
\examples{
\donttest{
cran_package_db()
}
}
\seealso{
cran_db_refresh
}
